% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ada_get_hostname}
\alias{ada_get_hostname}
\title{Get hostname component of URL}
\usage{
ada_get_hostname(url, decode = TRUE)
}
\arguments{
\item{url}{character. one or more URL to be parsed}

\item{decode}{logical. Whether to decode the output (see \code{\link[utils:URLencode]{utils::URLdecode()}}), default to \code{TRUE}}
}
\value{
logical
}
\description{
Get hostname component of URL
}
\examples{
ada_get_hostname("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
