% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{ada_has_credentials}
\alias{ada_has_credentials}
\title{Check if URL has credentials}
\usage{
ada_has_credentials(url)
}
\arguments{
\item{url}{character. URL to be parsed}
}
\value{
logical
}
\description{
Check if URL has credentials
}
\examples{
ada_has_credentials("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
