% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{ada_has_non_empty_password}
\alias{ada_has_non_empty_password}
\title{Check if URL has a non empty password}
\usage{
ada_has_non_empty_password(url)
}
\arguments{
\item{url}{character. URL to be parsed}
}
\value{
logical
}
\description{
Check if URL has a non empty password
}
\examples{
ada_has_non_empty_password("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
