% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{ada_has_non_empty_username}
\alias{ada_has_non_empty_username}
\title{Check if URL has a non empty username}
\usage{
ada_has_non_empty_username(url)
}
\arguments{
\item{url}{character. one or more URL to be parsed}
}
\value{
logical
}
\description{
Check if URL has a non empty username
}
\examples{
ada_has_non_empty_username("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
