% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{ada_has_search}
\alias{ada_has_search}
\title{Check if URL has a search}
\usage{
ada_has_search(url)
}
\arguments{
\item{url}{character. one or more URL to be parsed}
}
\value{
logical
}
\description{
Check if URL has a search
}
\examples{
ada_has_search("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
