% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{ada_url_parse}
\alias{ada_url_parse}
\title{Use ada-url to parse a url}
\usage{
ada_url_parse(url, decode = TRUE)
}
\arguments{
\item{url}{character. one or more URL to be parsed}

\item{decode}{logical. Whether to decode the output (see \code{\link[utils:URLdecode]{utils::URLdecode()}}), default to \code{TRUE}}
}
\value{
A data frame of the url components:
href, protocol, username, password, host, hostname, port, pathname, search, and hash
}
\description{
Use ada-url to parse a url
}
\details{
For details on the returned components refer to the introductory vignette.
}
\examples{
ada_url_parse("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
