% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labellength.R
\name{check_labellength}
\alias{check_labellength}
\title{ADA's standard label length check}
\usage{
check_labellength(label)
}
\arguments{
\item{label}{character vector; a vector containing all variable labels from a dataset. Can be retrieved using \link[sjlabelled]{get_label}}
}
\value{
Returns a filtered tibble with strings over the 255 word limit, and their length count.
}
\description{
Checks a character vector for any rows that contain >= 255 characters.
}
