% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{url_decode}
\alias{url_decode}
\title{Function to percent-decode characters in URLs}
\usage{
url_decode(url)
}
\arguments{
\item{url}{a character vector}
}
\description{
Similar to \link[utils:URLencode]{utils::URLdecode}
}
\examples{
url_decode("Hello\%20World")
}
