% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{url_decode2}
\alias{url_decode2}
\title{Function to percent-decode characters in URLs}
\usage{
url_decode2(url)
}
\arguments{
\item{url}{a character vector}
}
\value{
percent decoded URLs as character vector
}
\description{
Similar to \link[utils:URLdecode]{utils::URLdecode}
}
\details{
A \verb{\%} that is not followed by two hexadecimal digits is not a valid
escape sequence and is returned unchanged.
}
\examples{
url_decode2("Hello\%20World")
}
