% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{ada_url_parse}
\alias{ada_url_parse}
\title{Use ada-url to parse a url}
\usage{
ada_url_parse(url)
}
\arguments{
\item{url}{url}
}
\value{
A list of components
}
\description{
Use ada-url to parse a url
}
\examples{
ada_url_parse("https://user_1:password_1@example.org:8080/dir/../api?q=1#frag")
}
