% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{public_suffix}
\alias{public_suffix}
\title{Extract the public suffix from a vector of domains or hostnames}
\usage{
public_suffix(domains, icann_only = FALSE)
}
\arguments{
\item{domains}{character. vector of domains or hostnames}

\item{icann_only}{logical. Whether to use only the ICANN section of the
public suffix list, ignoring privately registered suffixes such as
\code{github.io} or \code{blogspot.com}. Defaults to \code{FALSE}.}
}
\value{
public suffixes of domains as character vector
}
\description{
Extract the public suffix from a vector of domains or hostnames
}
\details{
\code{domains} may be either full URLs or bare hostnames; anything that
does not parse as a URL is treated as a hostname. Wildcard (\code{*}) and
exception (\code{!}) rules of the public suffix list are both honoured.
}
\examples{
public_suffix("http://example.com")

# hostnames work too
public_suffix("example.com")

# for general URLs the hostname is extracted first
public_suffix("http://example.com/path/to/file")

# *.kobe.jp is a wildcard rule, !city.kobe.jp an exception to it
public_suffix(c("foo.kobe.jp", "city.kobe.jp"))

# privately registered suffixes are included by default
public_suffix("foo.github.io")
public_suffix("foo.github.io", icann_only = TRUE)
}
