% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{public_suffix}
\alias{public_suffix}
\title{Extract the public suffix from a vector of domains}
\usage{
public_suffix(domains)
}
\arguments{
\item{domains}{character. vector of domains}
}
\description{
Extract the public suffix from a vector of domains
}
