% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{public_suffix}
\alias{public_suffix}
\title{Extract the public suffix from a vector of domains}
\usage{
public_suffix(url)
}
\arguments{
\item{url}{character. one or more URL to be parsed}
}
\description{
Extract the public suffix from a vector of domains
}
