% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl.R
\name{public_suffix}
\alias{public_suffix}
\title{Extract the public suffix from a vector of domains or hostnames}
\usage{
public_suffix(domains)
}
\arguments{
\item{domains}{character. vector of domains or hostnames}
}
\value{
public suffixes of domains as character vector
}
\description{
Extract the public suffix from a vector of domains or hostnames
}
\examples{
public_suffix("http://example.com")

# doesn't work for general URLs
public_suffix("http://example.com/path/to/file")

# extracting hostname first does the trick
public_suffix(ada_get_hostname("http://example.com/path/to/file"))
}
