% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{url_decode2}
\alias{url_decode2}
\title{Function to percent-decode characters in URLs}
\usage{
url_decode2(url)
}
\arguments{
\item{url}{a character vector}
}
\description{
Similar to \link[utils:URLencode]{utils::URLdecode} and \code{\link[urltools:encoder]{urltools::url_decode()}}
}
\examples{
url_decode2("Hello\%20World")
}
